made changes to bot join event and image generation

+++ modules/image_generation.py
@@ -86,54 +86,57 @@ async def aiterms(ctx):
 @commands.command(name="ai")
 @commands.cooldown(1, 30, commands.BucketType.user)
 async def ai(ctx, *, msg):
-    save(ctx, msg)
-    try:
-        with open("ai_allowed.txt", "r") as read:
-            allowed_users = read.readlines()
-        allowed_users = [int(user_id.strip()) for user_id in allowed_users]  
-        if ctx.author.id not in allowed_users:
-            embed = discord.Embed(title="Error", description=f"Hi {ctx.author.name}, use ``!aiterms`` first to be authorized ")
-            embed.set_footer(text="Support Server: https://discord.gg/cvNa9XTbD9")
-            await ctx.send(embed=embed)
-            return
-        
-       
-        if ai.locked:
-            await ctx.reply(f"Hi {ctx.author.name} , The command is currently being used by someone else. Please wait for it to be available.")
-            return
-
-        ai.locked = True
-        ret = await ctx.send("Generating Image <a:loading:1108012790783946772>")
-        
-        API_URL = "https://api-inference.huggingface.co/models/andite/anything-v4.0"
-        headers = {"Authorization": HUGGING_FACE_KEY}
-
-        def query(payload):
-            response = requests.post(API_URL, headers=headers, json=payload)
-            return response.content
-
-        image_bytes = query({
-            "inputs": msg,
-        })
-
-        image = Image.open(io.BytesIO(image_bytes))
-
-        image.save("output.png")
-
-        with open('output.png', 'rb') as f:
-            picture = discord.File(f)
-            embed = discord.Embed(title="Generated Image", description=f"Prompt: {msg}", color=0x000000)
-            embed.set_image(url="attachment://output.png")
-            embed.set_footer(text="Note: Generating Explicit Images Will Result In A Ban")
-            await ret.delete()
-            await ctx.reply(embed=embed, file=picture)
+    if msg:
+        save(ctx, msg)
+        try:
+            with open("ai_allowed.txt", "r") as read:
+                allowed_users = read.readlines()
+            allowed_users = [int(user_id.strip()) for user_id in allowed_users]  
+            if ctx.author.id not in allowed_users:
+                embed = discord.Embed(title="Error", description=f"Hi {ctx.author.name}, use ``!aiterms`` first to be authorized ")
+                embed.set_footer(text="Support Server: https://discord.gg/cvNa9XTbD9")
+                await ctx.send(embed=embed)
+                return
             
-    except Exception as err:
-        await ctx.send(err)
+        
+            if ai.locked:
+                await ctx.reply(f"Hi {ctx.author.name} , The command is currently being used by someone else. Please wait for it to be available.")
+                return
 
-    finally:
-        ai.locked = False
-        os.remove('output.png')
+            ai.locked = True
+            ret = await ctx.send("Generating Image <a:loading:1108012790783946772>")
+            
+            API_URL = "https://api-inference.huggingface.co/models/andite/anything-v4.0"
+            headers = {"Authorization": HUGGING_FACE_KEY}
+
+            def query(payload):
+                response = requests.post(API_URL, headers=headers, json=payload)
+                return response.content
+
+            image_bytes = query({
+                "inputs": msg,
+            })
+
+            image = Image.open(io.BytesIO(image_bytes))
+
+            image.save("output.png")
+
+            with open('output.png', 'rb') as f:
+                picture = discord.File(f)
+                embed = discord.Embed(title="Generated Image", description=f"Prompt: {msg}", color=0x000000)
+                embed.set_image(url="attachment://output.png")
+                embed.set_footer(text="Note: Generating Explicit Images Will Result In A Ban")
+                await ret.delete()
+                await ctx.reply(embed=embed, file=picture)
+                
+        except Exception as err:
+            await ctx.send(err)
+
+        finally:
+            ai.locked = False
+            os.remove('output.png')
+    else:
+        await ctx.send("``Please Enter A Prompt!``")
 
 
 # Cooldown error handler
@@ -142,6 +145,9 @@ async def ai_error(ctx, error):
     if isinstance(error, commands.CommandOnCooldown):
         remaining = round(error.retry_after)
         await ctx.send(f"This command is on cooldown. Please try again in {remaining} second(s).")
+    else:
+         await ctx.send("``Please Enter A Prompt!``")
+
